--- backend/src/main/java/com/pickgo/domain/auth/dto/CreateTokenResponse.java
@@ -1,7 +1,0 @@
-package com.pickgo.domain.auth.dto;
-
-public record CreateTokenResponse(String accessToken) {
-	public static CreateTokenResponse of(String accessToken) {
-		return new CreateTokenResponse(accessToken);
-	}
-}
