+++ backend/src/main/java/com/pickgo/domain/auth/dto/TokenDetailResponse.java
@@ -1,0 +1,7 @@
+package com.pickgo.domain.auth.dto;
+
+public record TokenDetailResponse(String accessToken) {
+	public static TokenDetailResponse of(String accessToken) {
+		return new TokenDetailResponse(accessToken);
+	}
+}
