+++ backend/src/main/java/com/pickgo/domain/member/dto/MemberPrincipal.java
@@ -1,4 +1,4 @@
-package com.pickgo.domain.auth.dto;
+package com.pickgo.domain.member.dto;
 
 import java.util.UUID;
